Added a working 16x16 grid-box
Added a 16x16 grid-box within a container.
Added a generate button(not yet implemented) which should prompt for number of boxes per side.
With that, it should generate a new canvas (e.g 32 box per side... 32x32 grid = 1024 boxes).

Mainly `flex-basis: 6%;` and `flex-grow: 1;` does the job.

As for the listeners, I've added both for mouseover and mouseout.

whenever a mouse moves into a grid-box the background color is applied.
Also, when a mouse leaves the grid-box the background color is also applied.

Which clearly shows a trailing/hovering effect.
At first, I tried it in the css, using the `:hover` pseudo-class.