Refactor the process of creating the boxes
Based from my last commit, I've managed to create a `16x16` grid box but the button is not implemented yet.

As of now, the generate button is working as intended, whenever a user wants to generate a new canvas, it prompts the user, asking
for the boxes per side. Then it checks if a grid-boxes are already existing, if it is, it clears it and generate a new one based on the input.

As for the event listeners, I've temporarily moved them inside a `createBox()`. I had a hard time making the `mouseover` and
`mouseout` event listeners work. As the boxes are created, somehow the listeners are already listening for a non-existing class which is the
`.color-box`. This is based only my assumption, which is why I tried to move the event listeners right after the boxes are created.

I noticed that, by doing that, event listeners we're able to recognize the `.color-box` class created after clicking the generate button.

But this approach is ugly for me, I want my `createBox()` not to be chunky in code.

I was thinking that maybe I can toggle the css class for `.color-box` when the event listeners are triggered.