Add code experiment
Added some random code from the internet, I will be basing here some of
my works in the main branch.

The key differences from our code is that, I tried to create the all the
divs based on the user input, but the `flex-basis` is not changing.

Which resulted in some weird sized divs that are not equally matching
the desired outcome.