Refactor code and the process of creating divs
Instead of changing the `flex-basis` value based on the user input and
container size.

A nested for loop is used that creates actual row of boxes which will
serve as containers for the individual boxes present in a row.

Also, upon generating new canvas, the `#container` will invoke
`replaceChildren()` which will replace or remove(if no arguments passed)
the children and sub-children of the parent node which is the
`#container`.

Planning to add some feature in the next commit, where randomized rgb()
values is used in changeColors() instead of hardcoding the colors for
every `mouseover` event.